requirements.txt
transformers>=4.35.0
torch>=2.0.0
pandas>=2.0.0
numpy>=1.20.0
requests>=2.25.0
tqdm>=4.60.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
yfinance>=0.2.0
python-dotenv>=0.19.0
scikit-learn>=1.0.0
vaderSentiment>=3.3.0
newsapi-python>=0.2.0
finnhub-python>=2.4.0
twelvedata>=1.0.0
streamlit>=1.28.0
alpaca-trade-api>=3.0.0 